+++ ocabox-common_bases.py
@@ -3,7 +3,7 @@ __copyright__ = ""
 __credits__ = ["Ernest Lasek (elasek@akond.com)", "Mikołaj Kałuszyński (mkalusz@camk.edu.pl)"]
 __license__ = "MIT"
 # When bumping the version, also update: pyproject.toml [tool.poetry] version
-__version_num__ = [1, 2, 1]
+__version_num__ = [1, 2, 2]
 __version__ = "{}.{}.{}".format(*__version_num__)
 __maintainer__ = "Mikołaj Kałuszyński (mkalusz@camk.edu.pl)"
 __email__ = "mkalusz@camk.edu.pl"